flask_backend/requirements.txt
crewai == 0.121.1
crewai-tools == 0.45.0
python-dotenv == 1.1.0
pydantic == 2.11.4
Flask == 3.1.1
SQLAlchemy == 2.0.41
flask-cors == 6.0.0
Flask-JWT-Extended == 4.7.1
bcrypt == 4.3.0
Flask-SocketIO == 5.5.1
langchain == 0.3.25
langchain-core == 0.3.62
langgraph == 0.4.8
chromadb == 0.5.23